+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/dynamic/topdynamic_comment/DynamicCommentTopPresenter.java
@@ -76,8 +76,11 @@ public class DynamicCommentTopPresenter extends AppBasePresenter<DynamicCommentT
         addSubscrebe(subscription);
     }
 
+    /**
+     * @return 积分余额
+     */
     @Override
-    public double getBalance() {
+    public long getBalance() {
         AuthBean authBean = AppApplication.getmCurrentLoginAuth();
         if (authBean != null) {
             UserInfoBean userInfoBean = mUserInfoBeanGreenDao.getSingleDataFromCache(authBean.getUser_id());
